+++ src/webflickr.c
@@ -167,7 +167,7 @@ event_flickr_access_keys_generate (GtkWidget **gw_array)
 /**
  * @brief  Create file name using image title (display name) and url.
  *
- * @param[in]  s_disp_name  Diaplay name (image title)
+ * @param[in]  s_disp_name  Display name (image title)
  * @param[in]  s_image_url  Image url (used for getting extension)
  * @param[in]  s_image_id   Image id
  * @return     New string with name for saving file. After use it should be