+++ src/webwidget_c.c
@@ -39,10 +39,8 @@ const char *
 ww_name (const int i_site)
 {
     switch (i_site) {
-        case WEB_SERV_PEXELS:
-            return "Pexels";
-        case WEB_SERV_PIXBAY:
-            return "Pixbay";
+        case WEB_SERV_PIXABAY:
+            return "Pixabay";
         case WEB_SERV_WALLHAVEN:
             return "Wallhaven";
         case WEB_SERV_WALLABYSS:
@@ -64,10 +62,8 @@ const char *
 ww_opts (const int i_site)
 {
     switch (i_site) {
-        case WEB_SERV_PEXELS:
-            return "Pexels_opts";
-        case WEB_SERV_PIXBAY:
-            return "Pixbay_opts";
+        case WEB_SERV_PIXABAY:
+            return "Pixabay_opts";
         case WEB_SERV_WALLHAVEN:
             return "Wallhaven_opts";
         case WEB_SERV_WALLABYSS:
@@ -89,10 +85,8 @@ int
 ww_logo_id (const int i_site)
 {
     switch (i_site) {
-        case WEB_SERV_PEXELS:
-            return W_LOGO_PEXELS;
-        case WEB_SERV_PIXBAY:
-            return W_LOGO_PIXBAY;
+        case WEB_SERV_PIXABAY:
+            return W_LOGO_PIXABAY;
         case WEB_SERV_WALLHAVEN:
             return W_LOGO_WALLHAVEN;
         case WEB_SERV_WALLABYSS:
@@ -121,11 +115,8 @@ ww_get_api_key_names (const int i_site)
     ns_data = nstrings_new (4, NS_VAL_EMPTY);
 
     switch (i_site) {
-        case WEB_SERV_PEXELS:
-            str_append (&ns_data->s_str[0], "Pexels api");
-            break;
-        case WEB_SERV_PIXBAY:
-            str_append (&ns_data->s_str[0], "Pixbay api");
+        case WEB_SERV_PIXABAY:
+            str_append (&ns_data->s_str[0], "Pixabay api");
             break;
         case WEB_SERV_WALLHAVEN:
             str_append (&ns_data->s_str[0], "Wallhaven api");
@@ -457,4 +448,3 @@ search_opts_to_str (const Setting *st_setts)
     return s_res;
 }
 /*----------------------------------------------------------------------------*/
-